requirements.txt
streamlit==1.8.1
Pillow==7.0.0